List all servers

Lists all Beakr servers currently running (and any other servers created with the httpuv package). This function is included to encourage experimentation so that users who create multiple Beakr instances can quickly find and stop them all.

See httpuv::listServers for details.

listServers()

Value

None

Examples

library(beakr)

beakr1 <- newBeakr()
beakr2 <- newBeakr()
beakr1 %>% listen(daemon = TRUE, port = 1234, verbose = TRUE)
#> Beakr Instance: House Hollywood 
#>  State: Active | Host: 127.0.0.1 | Port: 1234 | Middlewares: 0 
beakr2 %>% listen(daemon = TRUE, port = 4321, verbose = TRUE)
#> Beakr Instance: Brother Moms 
#>  State: Active | Host: 127.0.0.1 | Port: 4321 | Middlewares: 0 
length(listServers())
#> [1] 2
stopAllServers()
length(listServers())
#> [1] 0