Listen for connections on a beakr instance

Binds and listens for connections on the specified host and port.

listen(beakr, host, port, daemon)

Arguments

beakr
a beakr instance.

host
a string that is a valid IPv4 or IPv6 address to listen on. Defaults to the local host ("127.0.0.1").

port
a number or integer that indicates the port to listen on. Default is a port opened on 8080.

daemon
run the instance in the background, the default is FALSE.

Value

A `beakr` App object.

Details

listen() binds the port and listens for connections on a thread. The thread handles the I/O, and when it receives a HTTP request, it will schedule a call to the user-defined middleware and handle the request.

If the daemon boolean value is set to TRUE, listen() binds the specified port and listens for connections on a thread running in the background.

See the httpuv package for more information.

Examples

if (FALSE) {
# Run in foreground
new_beakr() %>%
  http_get("/", function(req, res, err) {
    return("Successful GET request!\n")
  }) %>%
  listen()

# Run in background
#' new_beakr() %>%
  http_get("/", function(req, res, err) {
    return("Successful GET request!\n")
  }) %>%
  listen(daemon = TRUE)
}