Add CORS-headers middleware

Add Cross-Origin Resource Sharing (CORS) middleware to the instance.

cors(beakr, path = NULL, with_methods = c("POST", "GET", "PUT",
  "OPTIONS", "DELETE", "PATCH"), with_origin = "*",
  with_headers = NULL, with_credentials = NULL, max_age = NULL,
  expose_headers = NULL)

Arguments

beakr
a beakr instance.

path
a string path.

with_methods
tbd

with_origin
tbd

with_headers
tbd

with_credentials
tbd

max_age
tbd

expose_headers
tbd

Details

CORS is a mechanism that uses additional HTTP headers to tell browsers to give a web application running at one origin, access to selected resources from a different origin. A web application executes a cross-origin HTTP request when it requests a resource that has a different origin (domain, protocol, or port) from its own.

Additional Info: https://developer.mozilla.org/en-US/docs/Web/HTTP/Access_control_CORS